refactor: replace Double with BigDecimal in monetary and odds fields

+++ moabet/src/main/java/com/api/moabet/dto/bet/BetResponseDTO.java
@@ -1,14 +1,15 @@
 package com.api.moabet.dto.bet;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.api.moabet.models.enums.StatusBet;
 
 public record BetResponseDTO(
-        Long idBet,
-        Double amount,
-        StatusBet status,
-        LocalDateTime createdAt,
-        Long userId,
-        Long eventId) {
+                Long idBet,
+                BigDecimal amount,
+                StatusBet status,
+                LocalDateTime createdAt,
+                Long userId,
+                Long eventId) {
 }
