+++ moabet/src/main/java/com/api/moabet/dto/event/EventResponseDTO.java
@@ -1,13 +1,15 @@
 package com.api.moabet.dto.event;
 
+import java.math.BigDecimal;
+
 import com.api.moabet.models.enums.Result;
 import com.api.moabet.models.enums.StatusEvent;
 
 public record EventResponseDTO(
-        String name,
-        String description,
-        Double odds,
-        StatusEvent status,
-        Result result) {
+                String name,
+                String description,
+                BigDecimal odds,
+                StatusEvent status,
+                Result result) {
 
 }
