+++ moabet/src/main/java/com/api/moabet/dto/transaction/DepositResponseDTO.java
@@ -1,12 +1,13 @@
 package com.api.moabet.dto.transaction;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.api.moabet.models.enums.Type;
 
 public record DepositResponseDTO(
-                Double amount,
-                Long transactionId,
-                LocalDateTime createdAt,
-                Type type) {
+        BigDecimal amount,
+        Long transactionId,
+        LocalDateTime createdAt,
+        Type type) {
 }
